FIX: dated episode is identified as series
Sometimes, an episode is identified  as series, even when it has a date in the filename. 
This causes wrong filename to be created.

In case of a movie, there is a check if it contains a date, and if so it's marked as dated.
This check doesn't exists in case of an episode, so this fix adds it.