FIX: Videosort crushes with series name with 2 numbers and a dash (#17)
Videosort crushes with series name with 2 numbers and a dash.
This is caused due to a list is returned from guessit (and not string) when such a case occur.
When a list is retuned then a string is expected, Videosort crushes with a TypeError error.

Since no one is maintaining guessit on github anymore, there is no point waiting for a fix from them,
I have created a workaround (similar to other guessit workarounds), that joins all the parts of the list into one string.
I have also added a test case to test it in future times.